% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_explorer.R
\name{foldExplorer}
\alias{foldExplorer}
\title{Explore the generated folds}
\usage{
foldExplorer(blocks, rasterLayer, speciesData)
}
\description{
This function is deprecated! Please use \code{\link{cv_plot}} function for plotting the folds.
}
