% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explorer.R
\name{foldExplorer}
\alias{foldExplorer}
\title{Explore the generated folds}
\usage{
foldExplorer(blocks, rasterLayer, speciesData)
}
\arguments{
\item{blocks}{deprecated!}

\item{rasterLayer}{deprecated!}

\item{speciesData}{deprecated!}
}
\description{
This function is deprecated! Please use \code{\link{cv_plot}} function for plotting the folds.
}
