% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explorer.R
\name{rangeExplorer}
\alias{rangeExplorer}
\title{Explore spatial block size}
\usage{
rangeExplorer(
  rasterLayer,
  speciesData = NULL,
  species = NULL,
  rangeTable = NULL,
  minRange = NULL,
  maxRange = NULL
)
}
\arguments{
\item{rasterLayer}{raster layer for make plot}

\item{speciesData}{a simple features (sf) or SpatialPoints object containing species data (response variable). If provided, the species data will be shown on the map.}

\item{species}{character value indicating the name of the field in which the species data (response variable e.g. 0s and 1s) are stored.
If provided, species presence and absence data will be shown in different colours.}

\item{rangeTable}{deprecated option!}

\item{minRange}{a numeric value to set the minimum possible range for creating spatial blocks. It is used to limit the searching domain of
spatial block size.}

\item{maxRange}{a numeric value to set the maximum possible range for creating spatial blocks. It is used to limit the searching
domain of spatial block size.}
}
\description{
This function is deprecated and will be removed in future updates! Please use \code{\link{cv_block_size}} instead!
}
\seealso{
\code{\link{cv_block_size}}
}
